+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - The timeseries prop has been removed from bar charts, in favour of using that label handling by default (when there is not enough room for all diagonal labels, some will be dropped).
 
+### Added
+
+- `emptyStateText` and empty state handling to `<MultiSeriesBarChart />`
+
 ## [0.8.1] — 2021-04-20
 
 ### Added